Upgrade Emmet helper to latest version, add Emmet support inside components
