+++ packages/language-server/package.json
@@ -25,7 +25,7 @@
     "ts-morph": "^12.0.0",
     "typescript": "^4.5.4",
     "vscode-css-languageservice": "^5.1.1",
-    "vscode-emmet-helper": "2.1.2",
+    "@vscode/emmet-helper": "^2.8.4",
     "vscode-html-languageservice": "^3.0.3",
     "vscode-languageserver": "6.1.1",
     "vscode-languageserver-protocol": "^3.16.0",
@@ -35,7 +35,7 @@
   },
   "devDependencies": {
     "@types/lodash": "^4.14.116",
-    "astro": "0.23.5",
+    "astro": "^0.23.5",
     "astro-scripts": "0.0.1",
     "tap": "^15.0.9"
   }
